Moving files so they work both with Python2 and 3
