+++ masonry/cli.py
@@ -1,15 +1,18 @@
 import click
 import os
 import logging
+import sys
 
-from src.renderers import certifications_to_gitbook
+if sys.version_info[0] < 3:
+    reload(sys)
+    sys.setdefaultencoding('utf-8')
 
 from masonry.certification_builder import CertificationBuilder
 from masonry.inventory_builder import InventoryBuilder
 from masonry.gitbook_builder import GitbookBuilder
 
-from src import template_generator
-from src import utils
+from masonry.helpers import template_generator
+from masonry.helpers import utils
 
 
 def verify_certification_path(certification, certs_path):
@@ -91,9 +94,6 @@ def docs(export_format, certification, exports_dir, data_dir, output_dir):
             gitbook_output_dir = os.path.join(output_dir, 'gitbook')
             gitbook_markdown_dir = os.path.join(markdown_dir, 'gitbook')
             utils.create_dir(os.path.join(gitbook_output_dir, 'content'))
-            # output_path = certifications_to_gitbook.create_gitbook_documentation(
-            #   cert_path, gitbook_output_dir, gitbook_markdown_dir
-            # )
             output_path = GitbookBuilder(certification_yaml_path=cert_path).export(
                 gitbook_output_dir, gitbook_markdown_dir
             )
