+++ masonry/helpers/templates/components/system_name/component-name.yaml
